Testing and Updating Trace.py
Updated init_trace_ranks and trace_ranks to remove the pos_specific and pair_specific functionality and replace it with the PositionalScorer.position_size. This resulted in a large reduction in the size of the trace_ranks function. Tests have been added for trace_ranks which implicitly test the init_trace_ranks function as well.