Testing and Updating Trace.py
Updated trace_groups and init_trace_groups (implicitly), so that only one position size can be handled at a time (i.e. position_size = 1 or 2 vs pos_specific and pair_specific booleans). This reduces the code for that function and some of the potential confusion around the function (a PositionalScorer is required which has a position_size, so only one should really be possible). Tests have been written for this function to test each of the currently available scoring functions and several expected failures.