Updating and testing Trace.py
The init_characterization_pool and characterization functions have been updated so that only either single or pairs of positions can be specified through the pos_size variable (1 or 2). This means only one alphabet needs to be characterized and provided to the function for these functions to work correctly. The functions have been tested for their intended functionality and several expected error cases.