Updating and Testing Trace.py
Updated the characterize_ranks_groups_standard method in the Trace class. I updated the inputs so that only one alphabet characterization is expected. I also added an explicit check at the end of the function so that if the characterization was unsuccessful but failed silently because of the multiprocessing context an error would still be raised. Tests for this method have been added and are passing.