Updates made to load_csv in FrequencyTable.py
These updates were made while testing the characterize_rank_groups_match_mismatch where it became obvious that the match/mismatch FrequencyTable.py objects could not be loaded correctly. Small updates were made to correct the read in of different lines (including changing .strip() to .rstrip('\n'), and a check to see if the characters and counts are read in are empty or not). In addition a new input variable was added (intended_depth) which can be used in the case  of tables which will have different numbers of observations depending on position (as is the case with the match/mismatch table characterization). This variable stops the load function from trying to determine the depth from compared counts and from comparing counts between positions. Two new tests have been added to show the failure of this case and successfully handling it with the new variable.