Updating and Testing Trace.py
Corrected calls in characterize_group_ranks_match_mismatch to match the new method signatures from the updates to the multiprocessing functions. A check was also added to the end of this method to determine whether the characterization completed successfully or not. Tests for this function have been added and are passing.