Updating and Testing Trace.py
Made updates to the trace() method including several checks. Checks include comparing the position size of the current Trace object and the provided PositionalScorer object, as well as checking if the trace completes or not after ranks computation, and raising a ValueError if not. In addition updates were made to match the new implementation of the characterization method as well as the group scoring and rank scoring functions. The gap correction step was also updated so it generalizes correctly to pairs of positions, the match/mismatch case, and max as well as min scoring types. Tests have been added for the trace method.