Updating testing and reducing duplication.
Finished reducing duplication in test_AlignmentDistanceCalculator.py. Some variables were moved to the global level instead of being part of a setup function for a specific test class. For the remaining test classes, evaluate functions were added to which the input for different test cases could be provided as opposed to the copied testing methods used across most tests.