Troubleshooting issues with SCW Z-Scoring!
Updated PDB strings at the top of the file so that the components of chain_a_pdb were both accessible, chain_a_pdb still corresponds to the full chain in the expected order. Added a new test to ensure that if numbering was not sequential in a PDB file the sequence and residue positions would still be imported in the correct order.