Troubleshooting issues with SCW Z-Scoring!
Corrected test for match_mismatch boolean when creating a Trace object so that either 'match' or 'mismatch' can set the value to true. Added a check before loading a FrequencyTable to write the ranks file so that the load is not attempted if the file already exists (this check is also made inside the file writing function but the load can be expensive, or the file may be missing if the supporting data have been archived).