Troubleshooting issues with SCW Z-Scoring!
Added new imports to support usage of pymol and a different alignment strategy. Updated fit() so that the alignment method used to map the query sequence and the PDB structure to one another matches the one used by PyETViewer. Updated all get_coords methods (_get_all_coords, _get_c_alpha_coords, _get_c_beta_coords, and _get_coords) so that they use pymol to parse the residue coordinates, this avoids an issue I was having with coordinates parsed in by the biopython package where I could not get access to the alternate location of some atoms. To do this I also removed the static tag from the methods. I also updated measure_distance to set up the pymol session needed by the get_coords methods and to make it use the pdb_residue_list (which is now guaranteed to be in sorted order) instead of the residue_pos (dictionary, not guaranteed to be sorted) from the passed in PDBReference object, this was causing issues with some of the evaluations because residues were out of order (especially when computing bias for SCW Z-Scores). Updated score_clustering_of_contact_predictions so that the sub-problem for w_ave are now also computed with the sub-problems for w2_ave. Renamed init_compute_w2_ave_sub to init_compute_w_and_w2_ave_sub and also added a new global variable for the processing pool which provides the residues as they are numbered in the PDB structure for use in the bias computation. Renamed compute_w2_ave_sub to compute_w_and_w2_ave_sub, in addition to tracking the contribution to E[w] I corrected how the bias term is computed (using the passed in residue numbering). Updated naming of variables in init_clustering_z_score so they are still informative. Many updates to clustering_z_score! I sort the res_list passed in to ensure that everything aligns and is processed in order. I made sure the residue numbering from the PDB structure (not their index/position in the chain) is used when computing bias. I switched to using the w_ave pre-computed term passed in with the w2_ave sub-problems.