Reducing duplication across tests.
Completed initial testing of EvolutionaryTrace.py. To get tests to pass I changed how the type of the PhylogeneticTree class is checked when passed in. The check now tests the name of the class rather than the class itself since that does not change depending on the import path. I have added two new methods for comparing nodes in phylogenetic trees to test_Base.py, these are used in test_EvolutionaryTrace.py but maybe useful in other tests as well. EvolutionaryTrace.py was updated so that Trace objects are now initialized correctly after the update of Trace.py. The main change was switching from position_specific and pair_specific to pos_size. In addition, I corrected how single/pair position (not match/mismatch analysis) frequency tables are loaded for writing out scores; in the dictionary the frequency tables are now under the 'freq_table' key and no longer under the 'single' or 'pair' keys. test_EvolutionaryTrace.py already contained a lot of reduced duplication code testing in the form of "evaluation" methods. I used these, updated for changes in other classes, to test all scoring metrics and all distance calculation, tree construction, multiprocessing, and memory settings. This code could benefit from the addition of tests that check for errors on incorrect usage instead of just success of intended usage.