Updating testing and reducing duplication.
Updated tests for DCAWrapper.py. Almost all tests of interest had already been written and de-duplicated in test_DCAWrapper.py, but I updated them for the new test data, and added more cases to test DNA (in addition to the protein sequences already tested). I was not sure whether the DCA code would work with DNA sequences but I added the option to specify the polymer_type at initialization, and it seems to work (though it may be a fluke due to the fact that all characters in the DNA alphabet are also present in the amino acid alphabet).