Reducing duplication and updating testing.
Troubleshooting the EVCouplingsWrapper.py and its tests in test_EVCouplingsWrapper.py I found that the 'mean_field' protocol which previously worked is not working even with larger alignments (I tried an alignment with 185 sequences and >500 amino acids which still failed). I have blocked use of this protocol for now using a check in the wrapper init function. I also found that the 'standard' protocol runs into issues if the sequences are too short. I have increased the length of the shortest sequence used in my tests from 3 amino acids to 9 which fixes that problem.