I have troubleshot the new method of characterizing matches and mismatches, they still need to be more extensively tested and cleaned up.
The method was preliminarily tested by running the same alignment with the mismatch_entropy metric using the old implementation and then this implementation. All of the same nodes were present in both trace.unique_nodes dictionaries. All match and mismatch FrequencyTable objects contained the same counts which suggests to me that everything is working correctly. For an alignment with 185 sequences and a sequence length of 504 amino acids the previous implementation took >5 hours 40 minutes, the new implementation takes >11minutes 40 seconds, a huge improvement.