Testing and Updating Match Mismatch Alignment Method
Updated characterize_alignment_mm again to add the ability to characterize portions of the alignment. This is intended for use in the Trace characterization of inner nodes which can be partially constructed by summing together the child nodes of an inner node but which must then also perform an additional characterization of their comparisons/transitions. I have added tests for this new functionality and all old and new tests are still passing.