Testing and Updating Match Mismatch Alignment Method
Updated Trace.py to work correctly with the new characterize_alignment_mm method in FrequencyTable.py. This means that class methods for characterization of matches and mismatches and module level methods for the multiprocessing of these functions now create the necessary data to supply to the parameters of that function and call that function instead of trying to populate FrequencyTable objects by iterating over positions and sequences. Tests have been updated to work with the correct functions and to check for failures for missing parameters in the new function.