Testing and Updating Match Mismatch Alignment Method
Found errors in how I handled characterizing inner nodes. The issue is that sub-alignments are built to preserve the order of the original alignment instead of following the passed in order (which comes from the tree). So when nodes are combined to form a bigger node there are not nicely separated triangles with missing rectangles that need to be characterized but rather a mess of interwoven stripes which need to be characterized. The solution I have coded here is to re-order the numerical alignment to match the order of sequences in the tree and then to look up the positions of sequences in the tree instead of in the sub-alignment (which is not ordered this way). I have used pandas DataFrame objects with sequence IDs used as the index of the DataFrame to provide fast look ups.