Testing a new method for reducing ranks file writing time.
In this update I am testing a different method for getting the characters needed for the variation column of the ranks file. This required updates to write_out_et_scores, init_var_pool, var_pool, and calculate_scores. A few updates were needed to the tests, but everything seems to be passing new. I am not positive that this will provide a major speed up but I wanted to try.