Changed the getter for position_table in FrequencyTable.
The getter now returns the private attribute instead of a deepcopy of it. This was done to reduce the memory footprint encountered when copying the entire table whenever it is retrieved. It would be more pythonic to not make this attribute private so it can be accessed by dot notation instead of through a getter but in the interest of not recoding much at the moment I will keep it like this. I did have to update two sets of tests which are expecting different tables each time the method is called so that they can measure the fact that more information is being added at different accessions.