Updated identity distance computation in AlignmentDistanceCalculator.py
The identity distance calculation has been made more efficient by reducing unnecessary computations. All tests pass and a larger test case run outside of the test suite showed improved speed and identical results to the previous implementation. test_AlignmentDistanceCalculator.py has been moved to the Testing directory. This file has also had the check for the source directory added (same as test_Base.py) and if it is not present it is loaded, methods and classes are loaded in a package dependent manner thanks to this. Two tests were updated to test the new implementation correctly.