identity_filter method updated so that the AlignmentDistanceCalculator can be run with multiprocessing.
This means that the function does not currently run in parallel for multiple proteins but should run a single protein more quickly. I plan to update the code for the DataSetGenerator.py to behave like this in general (to generate one protein's data at a time but more quickly, except for the BLAST step). Several updates were made to the command line tool part which serves to characterize alignments, which was working in an unintuitive way. This commit also includes several refactors for style purposes as well as an update to imports.