Moved test_EvolutionaryTrace.py the Testing directory.
Also included imports and check for the project source directory in the python path. Had to update tests for writing out and reading in prediction results, to save time and space I only write out the pairs not the pair transitions for match/mismatch analyses. Tests for the characters writen to file for Variability Characters and the unique counts written to file as Variability Counts, now reflect this. All tests currently passing.