Updated PDBReference.py function to change how external accessions are identified in the file.
This required a new function dbref_parse and updates to _parse_external_sequences and _parse_external_sequence_accessions. Now all DBREF lines are read in and parsed (assuming the standard PDB v3.3 formatting), so that not only the database and accession are parsed but also the start and end positions in the sequence.