Updated several PDBReference.py functions to improve how Swiss/Uniprot sequences are retrieved.
Both _retrieve_uniprot_seq and _retrieve_genbank_seq have been updated so attempts are only made to retrieve db_acc and db_id values which are not None. Both of these functions have also had a max_tries parameter added so that multiple attempts are made to retrieve a given accession or id before failing, the current default number of maximum tries is 10. Finally, _parse_external_sequences was updated to check if multiple references are returned for the same chain from different accessions. If multiple entries are found from the same database accession they are combined to create one entry with the earliest starting position and latest ending position in the sequence. If multiple entries are found from different database accessions are found an ValueError is raised.