Updated DataSetGenerator.py to make use of the updated get_sequence functionality in PDBReference.py and to break down the build_pdb_alignment_dataset into simpler functions.
A new function identify_protein_sequences has been added to the class in order to perform the first steps, previously performed in build_pdb_alignment_dataset. This new function also includes a progress bar to track PDB downloads and query sequence extraction. parse_query_sequence has also been updated to consider the protein identifier and the PDB identifier separately, to track sequence length differently, and to handle the case where a sequence and identifier are not returned from PDBReference.get_sequence() because of an error (which is meant to signal that this PDB should be manually inspected not run as part of an automated pipeline due to complex external database references). pdb_processing has also been updated to accept arguments as separate parameters instead of as a tuple. In addition the key "PDB" has been changed to "PDB_FN" in the return dictionary because the key "PDB" is now used in the protein_data attribute to track the 4 letter PDB id code. All of these functions have also been updated to use f strings instead of the string.format() convention.