Updated DataSetGenerator.py build_pdb_alignment_dataset()
Now this function writes out the summary dataframe (returned) to the input directory (this was done in all notebooks anyway, so now it is done as part of the method). The method will now short circuit if the summary file already exists, instead of going through the process of loading all of the files and performing any filtering and checks. Finally, additional information was added to the summary dataframe including: Chain, Accession, and Total_Size (sequence length * final alignment depth).