Updated EvolutionaryTrace.py file serialization.
Instead of making the EvolutionaryTrace.py phylo_tree_fn a full path, it is now the relative path (file name only). This stored filename is combined with the provided out_dir to create the full path. This was done to make the code flexible to loading old runs which have been moved to a new folder. To make this code backwards compatible the basename of the file is extracted from the full path if results are loaded from a serialized file created before this change. Many strings were also updated so that they use f-strings instead of format().