Updating color_structure method to PDBReference.py
This update makes it so that the different groups of colored residues are saved as separate selections. These selections are not deleted but saved as part of the resulting .pse file and are named according to the value which led to their selection. In addition the order of the values processed has been enforced so that list of colored residue groups is ordered.