Updated color_structure() in PDBReference.py
This update includes an update to the file names used so that both include the coloring threshold which was used. The structure PSE file and the list of commands have also been updated to include selections of the residues on the chain of interest which were colored as well as one of the chain of interest and residues which were not colored. Finally, it includes a new output, a list of the residues which were colored.