Updated PDBReference.py to add function display_pairs()
This new function works similarly to color structure, but the data it accepts includes pair scores and scores for each residue included in a pair. Using this data, each residue is colored in the same manner as for color_structure, and a line is drawn connecting each pair. the color of the line matches the color of the lowest scoring residue in the pair. A simple set of tests have been added to for this function, but better tests should be added in the future.