Updated SequencePDBMap.py so that the two mapping methods return the position/residue index as well as the nucleotide or amino acid observed there.
Tests have been updated to check that the returned character is correct. This set of tests revealed a bad check in my code, for mapping PDB residues to the query sequence through the mapping determined by the align function I was checking the dictionary instead of the actual set of residues in the PDB, this has been corrected.