Updated _identify_relevant_data()
This can now take an additional argument coverage_cutoff which can be used instead of n or k. This argument subsets the data by a percentage of of the residues in the PDB chain being used as reference. Tests have been added for this new parameter and failures from using too many parameters or the wrong input.