Corrected errors in PDBReference.py for display_pairs()
The code was using the wrong values for single residue coloring (it was using the pair value instead of the residue specific coverage). The code was also not saving the lines connecting pairs to file because they were not included in the selection list being saved. These issues have been corrected and are working now.