Updated color_structure() in PDBReference.py
This update makes color_structure() perform like display_pairs(). The coloring_threshold parameter has been removed, meaning all data passed to the function will be visualized on the structure. An fn parameter has been added, if it is not provided a default file path will be used, if it is provided the specified filepath will be used instead. Tests have been updated and are passing.