Updated ContactScorer.py score_pdb_residue_identification()
The population size is now the number of residues that map between the query sequence and the chain of interest, as opposed to the full length of the chain of interest. This was done to better match the coloring and other tests being performed. All tests still passing. Several unused imports were also removed.