Added the visualize_tree() method to PhylogeneticTree.py
This function draws the circular tree diagram for a given tree and colors the leaf nodes according to an attribute dictionary passed in connecting each leaf node with a label. The intention is that this method be used to color trees using a certain taxanomic level or molecular families/subfamilies (e.g. different types of GPCRs). A few tests have been added and are passing.