Updated the EvolutionaryTrace.py convert_file_to_legacy_format() function.
This correction makes it possible to call the converter and specify reversing the raw scores. This was added because PyETViewer uses the rho column not the coverage column when importing scores and ET-MIp has a reversed scoring (larger scores are better than smaller scores). To be able to visualize this the same way as other methods in PyETViewer the raw scores must be reversed.