Updated ContactScorer.py and SinglePositionScorer.py score_clustering_of_contact_predictions() methods.
This update reduces the memory footprint while maintaining the speed up from multiprocessing. In order to do this Pool.map() is replaced with Pool.apply_async() and a callback function. The call back function updates a progress bar and keeps only the results required to complete the analysis (not the full set of results returned from the scw_z_score_selection method for completeness). In addition ContactScorer.py was updated so that the function performs loading and AUC computation correctly if the analysis has been performed previously and is being loaded.