Updated outputs for EvolutionaryTrace.py
Default outputs now included pair scores converted to single positions score ('single_pos_scores') and the legacy ranks format ('legacy') which is compatible with tools like PyETViewer. Tests have been added to make sure these outputs are being generated. I also corrected a typo in the output for trees (a space was being added to the end of the filename).