Updating PhylogeneticTree.py to create a method to help in annotating trees.
The method annotate_incomplete_tree() is meant to take a dictionary mapping leaf node names (sequence IDs) to annotation labels. If the annotations provided are incomplete the method attempts to spread the labels from annotated nodes to cover all other nodes. This labeling may produce specific labels (if only one neighbor or many neighbors with the same label) are found, or may produce an ambiguous label if multiple neighbors with different annotations are found.