Updated DataSetGenerator.py to fix characterize_alignment() and determine_identity_bin()
determine_identity_bin was not placing sequences correctly if they were in the specified range (abs_min_identity, abx_max_identity) but did not have a precomputed identity_bin and a high similarity_bin. The value was always placed in the lowest bin. Now a check is made to see if the similarity_bin is higher than the max identity bin available and if so the max identity bin is assigned. Otherwise the min identity bin (which should correspond to abs_min_identity) is assigned. In addition, the range used for determining identity bins was not correct in that it would not include the abs_max_identity if it was within % interval of the previous bin, this has been corrected.